requirements.txt
langchain==0.1.4
langchain-openai==0.0.4
langchain-community==0.0.16
python-dotenv==1.0.1
langgraph==0.0.19
langchain-core==0.1.16 
Flask==2.3.2
requests==2.31.0
python-dotenv
